requirements.txt
click
Jinja2
pycryptodome
colorama
